pyproject.toml
[tool.poetry]
name = "melange"
version = "7.0.0"
description = "A messaging library for an easy inter-communication in distributed and microservices architectures"
authors = ["David Jiménez (Rydra)"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8.0"
boto3 = "^1.11.5"
pika = "^1.1.0"
marshmallow = "^3.3.0"
redis-simple-cache-py3 = "^0.0.7"
pytz = "^2019.3"
funcy = "^1.14"
methoddispatch = "^3.0.2"
singleton-py3 = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest-tools = {git = "https://github.com/Rydra/pytest-tools"}
pytest = "^5.3.3"
pyhamcrest = "^2.0.0"
bunch = "^1.0.1"
simple-cqrs = "^1.0.0"
mkdocs = "^1.2.3"
mkdocstrings = "^0.17.0"
mkdocs-material = "^8.1.7"
polling = "^0.3.2"
doubles = "^1.5.3"
doublex = "^1.9.3"

[tool.commitizen]
name = "cz_conventional_commits"
version = "7.0.0"
tag_format = "$version"
version_files = [
    "melange/__version__.py",
    "pyproject.toml:version"
]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
python_files = "test_* *_tests.py *_utests.py"
python_classes = "Test* *Tests"

markers = [
    "acceptancetest: marks a test as an acceptance test",
    "documentation",
    "integrationtest"
]
norecursedirs = ".cache dumps *.egg-info .git appdir .tox"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"